Update the process to set the neo4j password
In neo4j 5.0 the syntax for setting the password using the
set-initial-password command has changed. This changes the password
setting task into a block that will run whichever command is
appropriate based on the version of the neo4j package that is installed
on the system. This also adds a `no_log: True` parameter to both of the
tasks to set the password so that the given password is not
accidentally leaked if there is an error.